requirements.txt
pyyaml>=6.0
openai>=1.0
PyGithub>=2.0
